refactor(provider): split StartCommand god-file by concern (1133->121 LOC)
Split the Start command struct into same-target extension files: core keeps
the command declaration, options, and run() orchestrator; serving modes
(standalone/foreground/scheduled), preflight + inline login, launchd daemon
install, interactive catalog picker, and the raw-mode TUI picker move to
StartCommand+Modes/+Preflight/+Daemon/+Picker/+TUIPicker.

Bodies byte-identical; 7 private methods reached across the split widened to
internal. Behavior unchanged.

Build green; swift test 1052/73 pass (unchanged from baseline).
